refactor: window components into window directory
