+++ src/components/window/MemoWindow.tsx
@@ -1,17 +1,23 @@
+import { useEditor } from "@tiptap/react";
+import StarterKit from "@tiptap/starter-kit";
 import { useEffect, useState } from "react";
-import type { BrowserWindowType } from "../types/desktop";
+import type { MemoWindowType } from "../../types/desktop";
+import TiptapEditor from "../TiptapEditor";
+import { Button } from "../ui/button";
 
-export function BrowserWindow({
+export function MemoWindow({
 	window,
 	onClose,
 	onMinimize,
+	onContentChange,
 	onBringToFront,
 	onPositionChange,
 	onSizeChange,
 }: {
-	window: BrowserWindowType;
+	window: MemoWindowType;
 	onClose: () => void;
 	onMinimize: () => void;
+	onContentChange: (content: string) => void;
 	onBringToFront: () => void;
 	onPositionChange: (position: { x: number; y: number }) => void;
 	onSizeChange: (size: { width: number; height: number }) => void;
@@ -24,6 +30,17 @@ export function BrowserWindow({
 		y: 0,
 		width: 0,
 		height: 0,
+	});
+
+	const editor = useEditor({
+		extensions: [StarterKit],
+		content: window.content,
+		editorProps: {
+			attributes: {
+				class:
+					"prose prose-sm prose-li:marker:text-black prose-p:m-0 prose-headings:m-0 prose-ul:m-0 prose-ol:m-0 prose-blockquote:m-0 prose-hr:m-0 prose-pre:m-0 m-5 focus:outline-none text-left",
+			},
+		},
 	});
 
 	const handleMouseDown = (e: React.MouseEvent) => {
@@ -60,8 +77,8 @@ export function BrowserWindow({
 					y: e.clientY - dragStart.y,
 				});
 			} else if (isResizing) {
-				const newWidth = Math.max(400, resizeStart.width + (e.clientX - resizeStart.x));
-				const newHeight = Math.max(300, resizeStart.height + (e.clientY - resizeStart.y));
+				const newWidth = Math.max(300, resizeStart.width + (e.clientX - resizeStart.x));
+				const newHeight = Math.max(200, resizeStart.height + (e.clientY - resizeStart.y));
 				onSizeChange({
 					width: newWidth,
 					height: newHeight,
@@ -85,9 +102,13 @@ export function BrowserWindow({
 		};
 	}, [isDragging, isResizing, dragStart, resizeStart, onPositionChange, onSizeChange]);
 
+	const handleSaveMemo = () => {
+		//todo: save to db
+		onClose();
+	};
 	return (
 		<div
-			className="fixed overflow-hidden rounded-lg border border-gray-200 bg-white shadow-2xl"
+			className="fixed flex flex-col overflow-hidden rounded-lg border border-gray-200 bg-white shadow-2xl"
 			style={{
 				left: window.position.x,
 				top: window.position.y,
@@ -99,7 +120,7 @@ export function BrowserWindow({
 			onMouseDown={handleMouseDown}
 		>
 			{/* Window Header */}
-			<div className="window-header flex cursor-grab items-center justify-between border-gray-200 border-b bg-gray-50 px-4 py-2 active:cursor-grabbing">
+			<div className="window-header flex flex-shrink-0 cursor-grab items-center justify-between border-gray-200 border-b bg-gray-50 px-4 py-2 active:cursor-grabbing">
 				<div className="flex items-center space-x-2">
 					<div className="flex space-x-2">
 						<button
@@ -119,21 +140,12 @@ export function BrowserWindow({
 					</div>
 					<span className="ml-4 font-medium text-gray-700 text-sm">{window.title}</span>
 				</div>
-				<div className="flex items-center space-x-2">
-					<div className="rounded bg-gray-100 px-2 py-1 text-gray-500 text-xs">{window.url}</div>
-				</div>
+				<Button onClick={handleSaveMemo}>Save</Button>
 			</div>
 
-			{/* Browser Content */}
-			<div className="h-full flex-1">
-				<iframe
-					src={window.url}
-					className="h-full w-full border-0"
-					style={{
-						height: "calc(100% - 50px)",
-					}}
-					title={window.title}
-				/>
+			{/* Window Content */}
+			<div className="min-h-0 flex-1 overflow-y-auto ">
+				<TiptapEditor editor={editor} />
 			</div>
 
 			{/* Resize Handle */}
